# docs/swagger-doc.yaml
openapi: 3.0.1
info:
  title: Tunemyles API
  description: >-
    This is a sample server Petstore server. You can find out more about Swagger
    at [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net](irc.freenode.net)
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'http://localhost:5000/api/v2'
    description: Dev Server API URL
...
paths:
...
  /account/merchant:
    post:
      tags:
        - account
      summary: Activate User Merchant (Route All Roles)
      description: Activate User Merchant (Route All Roles)
      operationId: activateMerchant
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ActivateMerchantBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuthorizedResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
  /account/me:
    get:
      tags:
        - account
      summary: Get Profile (Route All Roles)
      description: Get Profile (Route All Roles)
      operationId: getProfile
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProfileResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  /account/update-profile:
    patch:
      tags:
        - account
      summary: Update Profile (Route All Roles)
      description: Update Profile (Route All Roles)
      operationId: updateProfile
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateProfileBody'
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  /account/forgot-password:
    post:
      tags:
        - account
      summary: Forgot Password
      description: Forgot Password
      operationId: forgotPassword
      requestBody:
        description: Request Data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-codegen-request-body-name: body
  '/account/verify-token/{token}':
    patch:
      tags:
        - account
      summary: Verify Token
      description: Verify Token
      operationId: verifyToken
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /account/reset-password:
    patch:
      tags:
        - account
      summary: Reset Password
      description: Reset Password
      operationId: resetPassword
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        description: Request Data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      x-codegen-request-body-name: body
...
  '/banks/users/{id}':
    patch:
      tags:
        - banks
      summary: Update Bank User (Route All Roles)
      description: Update Bank User (Route All Roles)
      operationId: updateBankUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBankUserBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
    delete:
      tags:
        - banks
      summary: Delete Bank User (Route All Roles)
      description: Delete Bank User (Route All Roles)
      operationId: deleteBankUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: number
        success:
          type: boolean
        messages:
          type: string
    TokenResponse:
      type: object
      properties:
        token:
          type: string
    UserAuthorizedResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                type:
                  type: integer
                token:
                  type: string
...
    ActivateMerchantBody:
      required:
        - identity_photo
        - market_photo
        - no_identity
        - market_name
        - market_address
        - market_lat
        - market_lon
        - market_close_time
      properties:
        identity_photo:
          type: string
          format: binary
        market_photo:
          type: string
          format: binary
        no_identity:
          type: string
        market_name:
          type: string
        market_address:
          type: string
        market_lat:
          type: number
        market_lon:
          type: number
        market_close_time:
          type: string
...
    GetProfileResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                full_name:
                  type: string
                username:
                  type: string
                address:
                  type: string
                phone_number:
                  type: string
                img_url:
                  type: string
                  nullable: true
                wallets:
                  type: array
                  items:
                    type: object
                    properties:
                      balance:
                        type: integer
                      visibility:
                        type: integer
                      wallet_name:
                        type: string
                      wallet_description:
                        type: string
                      logo_label:
                        type: string
                        nullable: true
                      logo_url:
                        type: string
                        nullable: true
    UpdateProfileBody:
      type: object
      properties:
        avatar:
          type: string
          format: binary
        full_name:
          type: string
        address:
          type: string
        phone_number:
          type: string
    ForgotPasswordBody:
      required:
        - phone_number
      type: object
      properties:
        phone_number:
          type: string
    VerifyTokenResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TokenResponse'
    ResetPasswordBody:
      required:
        - new_password
      type: object
      properties:
        new_password:
          type: string
...
    UpdateBankUserBody:
      required:
        - bank_id
      type: object
      properties:
        bank_id:
          type: integer
        account_number:
          type: string
        account_name:
          type: string
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      name: authorization
      in: header